+++ config/routes.js
@@ -2,7 +2,7 @@ const posts = require('../actions/posts')
 const accounts = require('../actions/auth/accounts')
 const sessions = require('../actions/auth/sessions')
 
-module.exports = function (router) {
+module.exports = async (router) => {
   router
     .use(sessions.currentUser)
