+++ public/serviceWorker.js
@@ -1,0 +1,1 @@
+// nop
